[GUI 버그 수정] - Object 상속 객체의 소멸자가 호출이 안되서 메모리 누수되는 문제 수정 - GD32F4xx 관련 코드 일부 추가

+++ yss/gui/CheckBox.cpp
@@ -58,12 +58,12 @@ void CheckBox::paint(void)
 
 	clear();
 
-	drawRect(Pos{2, 2}, Size{width - 5, height - 5});
+	drawRect(Pos{2, 2}, Size{(unsigned short)(width - 5), (unsigned short)(height - 5)});
 
 	if (mState)
 	{
 		p1 = Pos{4, 4};
-		p2 = Pos{width / 2, height - 6};
+		p2 = Pos{(signed short)(width / 2), (signed short)(height - 6)};
 		drawLine(p1, p2);
 		p1.x++;
 		p2.x++;
@@ -71,7 +71,7 @@ void CheckBox::paint(void)
 
 		p1 = p2;
 		p1.x--;
-		p2 = Pos{width - 6, 4};
+		p2 = Pos{(signed short)(width - 6), 4};
 		drawLine(p1, p2);
 		p1.x++;
 		p2.x++;
